[Help] : Probando integración de Detalle Rol Usuario Requerimos documentación sobre mapping de tablas con llave compuesta y join de varias tablas

+++ Dogtor_Vet/src/main/java/com/veterinaria/controller/RazaController.java
@@ -17,7 +17,7 @@ public class RazaController {
 	
 	@RequestMapping("listaRaza")
 	@ResponseBody
-	public List<Raza> listaRaza(Long especie) {
+	public List<Raza> listaRaza(Integer especie) {
 		return service.listaRaza(especie);
 	}
 }
