+++ Dogtor_Vet/src/main/java/com/veterinaria/repository/RazaRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.repository.query.Param;
 
 import com.veterinaria.entity.Raza;
 
-public interface RazaRepository extends JpaRepository<Raza, Long>{
+public interface RazaRepository extends JpaRepository<Raza, Integer>{
 	
 	@Query("select r from Raza r where r.codigo_especie_mascota like :param_especie")
-	public abstract List<Raza> listaRaza(@Param("param_especie") Long especie);
+	public abstract List<Raza> listaRaza(@Param("param_especie") Integer especie);
 }
