+++ Dogtor_Vet/src/main/java/com/veterinaria/service/RazaService.java
@@ -5,5 +5,5 @@ import java.util.List;
 import com.veterinaria.entity.Raza;
 
 public interface RazaService {
-	public abstract List<Raza> listaRaza(Long especie);
+	public abstract List<Raza> listaRaza(Integer especie);
 }
